refactor: 'attachments' command handling
Move the 'attachments' command code and globals into a separate file.
The lists are now private to this file.

When an 'attachments' command is read, a notification will be sent.
