+++ Makefile.autosetup
@@ -62,14 +62,14 @@ default:	all
 ###############################################################################
 # neomutt
 NEOMUTT=	neomutt$(EXEEXT)
-NEOMUTTOBJS=	browser.o commands.o command_parse.o \
+NEOMUTTOBJS=	attachments.o browser.o commands.o command_parse.o \
 		complete.o conststrings.o context.o copy.o \
 		editmsg.o enriched.o enter.o flags.o functions.o git_ver.o \
 		handler.o hdrline.o help.o hook.o icommands.o init.o \
 		keymap.o mailcap.o maillist.o main.o menu.o muttlib.o mutt_account.o \
 		mutt_attach.o mutt_body.o mutt_commands.o mutt_config.o \
 		mutt_header.o mutt_history.o mutt_logging.o mutt_mailbox.o \
-		mutt_parse.o mutt_signal.o mutt_socket.o mutt_thread.o mx.o \
+		mutt_signal.o mutt_socket.o mutt_thread.o mx.o \
 		myvar.o opcodes.o postpone.o progress.o \
 		recvattach.o recvcmd.o resize.o rfc3676.o score.o \
 		sort.o state.o status.o subjectrx.o system.o version.o
