+++ commands.c
@@ -48,6 +48,7 @@
 #include "ncrypt/lib.h"
 #include "pager/lib.h"
 #include "send/lib.h"
+#include "attachments.h"
 #include "browser.h"
 #include "context.h"
 #include "copy.h"
@@ -62,7 +63,6 @@
 #include "mutt_logging.h"
 #include "mutt_mailbox.h"
 #include "mutt_menu.h"
-#include "mutt_parse.h"
 #include "muttlib.h"
 #include "mx.h"
 #include "options.h"
