+++ hdrline.c
@@ -46,11 +46,11 @@
 #include "gui/lib.h"
 #include "hdrline.h"
 #include "ncrypt/lib.h"
+#include "attachments.h"
 #include "format_flags.h"
 #include "hook.h"
 #include "maillist.h"
 #include "mutt_globals.h"
-#include "mutt_parse.h"
 #include "mutt_thread.h"
 #include "muttlib.h"
 #include "sort.h"
