+++ index/observer.c
@@ -33,6 +33,7 @@
 #include "email/lib.h"
 #include "core/lib.h"
 #include "gui/lib.h"
+#include "attachments.h"
 #include "context.h"
 #include "mutt_globals.h"
 #include "options.h"
@@ -198,6 +199,22 @@ static int index_subjrx_observer(struct NotifyCallback *nc)
 }
 
 /**
+ * index_attach_observer - Listen for attachment command changes affecting the Index/Pager - Implements ::observer_t
+ */
+static int index_attach_observer(struct NotifyCallback *nc)
+{
+  if (!nc->global_data)
+    return -1;
+  if (nc->event_type != NT_ATTACH)
+    return 0;
+
+  // struct MuttWindow *dlg = nc->global_data;
+
+  mutt_attachments_reset(ctx_mailbox(Context));
+  return 0;
+}
+
+/**
  * index_add_observers - Add Observers to the Index Dialog
  * @param dlg Index Dialog
  */
@@ -208,6 +225,7 @@ void index_add_observers(struct MuttWindow *dlg)
 
   notify_observer_add(NeoMutt->notify, NT_CONFIG, index_config_observer, dlg);
   notify_observer_add(NeoMutt->notify, NT_SUBJRX, index_subjrx_observer, dlg);
+  notify_observer_add(NeoMutt->notify, NT_ATTACH, index_attach_observer, dlg);
 }
 
 /**
@@ -221,4 +239,5 @@ void index_remove_observers(struct MuttWindow *dlg)
 
   notify_observer_remove(NeoMutt->notify, index_config_observer, dlg);
   notify_observer_remove(NeoMutt->notify, index_subjrx_observer, dlg);
+  notify_observer_remove(NeoMutt->notify, index_attach_observer, dlg);
 }
