+++ ncrypt/crypt.c
@@ -48,12 +48,12 @@
 #include "crypt.h"
 #include "lib.h"
 #include "send/lib.h"
+#include "attachments.h"
 #include "context.h"
 #include "copy.h"
 #include "cryptglue.h"
 #include "handler.h"
 #include "mutt_globals.h"
-#include "mutt_parse.h"
 #include "muttlib.h"
 #include "options.h"
 #include "state.h"
