+++ nntp/nntp.c
@@ -51,11 +51,11 @@
 #include "hcache/lib.h"
 #include "ncrypt/lib.h"
 #include "adata.h"
+#include "attachments.h"
 #include "edata.h"
 #include "hook.h"
 #include "mdata.h"
 #include "mutt_logging.h"
-#include "mutt_parse.h"
 #include "mutt_socket.h"
 #include "muttlib.h"
 #include "mx.h"
