+++ pattern/exec.c
@@ -46,10 +46,10 @@
 #include "lib.h"
 #include "ncrypt/lib.h"
 #include "send/lib.h"
+#include "attachments.h"
 #include "copy.h"
 #include "handler.h"
 #include "maillist.h"
-#include "mutt_parse.h"
 #include "muttlib.h"
 #include "mx.h"
 #include "state.h"
