+++ recvattach.c
@@ -43,6 +43,7 @@
 #include "recvattach.h"
 #include "ncrypt/lib.h"
 #include "send/lib.h"
+#include "attachments.h"
 #include "commands.h"
 #include "context.h"
 #include "format_flags.h"
@@ -54,7 +55,6 @@
 #include "mutt_attach.h"
 #include "mutt_globals.h"
 #include "mutt_menu.h"
-#include "mutt_parse.h"
 #include "mutt_thread.h"
 #include "muttlib.h"
 #include "mx.h"
