+++ send/send.c
@@ -49,6 +49,7 @@
 #include "compose/lib.h"
 #include "ncrypt/lib.h"
 #include "pattern/lib.h"
+#include "attachments.h"
 #include "browser.h"
 #include "context.h"
 #include "copy.h"
@@ -62,7 +63,6 @@
 #include "mutt_globals.h"
 #include "mutt_header.h"
 #include "mutt_logging.h"
-#include "mutt_parse.h"
 #include "muttlib.h"
 #include "options.h"
 #include "protos.h"
