+++ send/sendlib.c
@@ -48,12 +48,12 @@
 #include "sendlib.h"
 #include "lib.h"
 #include "ncrypt/lib.h"
+#include "attachments.h"
 #include "context.h"
 #include "copy.h"
 #include "handler.h"
 #include "mutt_globals.h"
 #include "mutt_mailbox.h"
-#include "mutt_parse.h"
 #include "muttlib.h"
 #include "mx.h"
 #include "options.h"
